Improve club management
- Move entities to /clubs/entities
- Add permissions (without CASL)
- Restrict roles to an enum, but add an additional label

+++ api/src/clubs/clubs.service.ts
@@ -1,16 +1,19 @@
 import { UniqueConstraintViolationException, wrap } from '@mikro-orm/core';
 import { InjectRepository } from '@mikro-orm/nestjs';
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { BadRequestException, ForbiddenException, Injectable } from '@nestjs/common';
 import { BaseRepository } from '../shared/lib/repositories/base.repository';
+import { ClubRole } from '../shared/lib/types/club-role.enum';
+import { Role } from '../shared/modules/authorization/types/role.enum';
 import type { PaginationOptions } from '../shared/modules/pagination/pagination-option.interface';
 import type { PaginatedResult } from '../shared/modules/pagination/pagination.interface';
 import { User } from '../users/user.entity';
-import { ClubMember } from './club-member.entity';
-import { Club } from './club.entity';
+import { ClubSearchService } from './club-search.service';
 import type { CreateClubMemberDto } from './dto/create-club-member.dto';
 import type { CreateClubDto } from './dto/create-club.dto';
 import type { UpdateClubMemberDto } from './dto/update-club-member.dto';
 import type { UpdateClubDto } from './dto/update-club.dto';
+import { ClubMember } from './entities/club-member.entity';
+import { Club } from './entities/club.entity';
 
 @Injectable()
 export class ClubsService {
@@ -18,41 +21,52 @@ export class ClubsService {
     @InjectRepository(Club) private readonly clubRepository: BaseRepository<Club>,
     @InjectRepository(ClubMember) private readonly clubMemberRepository: BaseRepository<ClubMember>,
     @InjectRepository(User) private readonly userRepository: BaseRepository<User>,
-    ) {}
+    private readonly clubSearchService: ClubSearchService,
+  ) {}
 
-  public async create(createClubDto: CreateClubDto): Promise<Club> {
+  public async create(user: User, createClubDto: CreateClubDto): Promise<Club> {
     const club = new Club(createClubDto);
 
     try {
       await this.clubRepository.persistAndFlush(club);
     } catch (error: unknown) {
       if (error instanceof UniqueConstraintViolationException)
-        throw new BadRequestException('Badge already exists');
+        throw new BadRequestException('Club already exists');
       throw error;
     }
+
+    const member = new ClubMember({ user, club, role: ClubRole.President });
+    await this.clubMemberRepository.persistAndFlush(member);
+    await this.clubSearchService.add(club);
 
     return club;
   }
 
   public async findAll(paginationOptions?: PaginationOptions): Promise<PaginatedResult<Club>> {
-    return await this.clubRepository.findWithPagination(paginationOptions);
+    return await this.clubRepository.findWithPagination(paginationOptions, {}, { populate: ['members', 'members.user'] });
   }
 
   public async findOne(clubId: number): Promise<Club> {
-    return await this.clubRepository.findOneOrFail({ clubId });
+    return await this.clubRepository.findOneOrFail({ clubId }, ['members', 'members.user']);
   }
 
-  public async update(clubId: number, updateClubDto: UpdateClubDto): Promise<Club> {
-    const club = await this.clubRepository.findOneOrFail({ clubId });
+  public async update(user: User, clubId: number, updateClubDto: UpdateClubDto): Promise<Club> {
+    const club = await this.clubRepository.findOneOrFail({ clubId }, ['members', 'members.user']);
+
+    // TODO: Move this to CASL
+    if (!user.roles.includes(Role.Admin) && !club.isClubAdmin(user))
+      throw new ForbiddenException('Not a club admin');
 
     wrap(club).assign(updateClubDto);
     await this.clubRepository.flush();
+    await this.clubSearchService.update(club);
     return club;
   }
 
   public async remove(clubId: number): Promise<void> {
     const club = await this.clubRepository.findOneOrFail({ clubId });
     await this.clubRepository.removeAndFlush(club);
+    await this.clubSearchService.remove(club.clubId.toString());
   }
 
   public async findAllUsersInClub(
@@ -60,15 +74,35 @@ export class ClubsService {
     paginationOptions?: PaginationOptions,
   ): Promise<PaginatedResult<ClubMember>> {
     const club = await this.clubRepository.findOneOrFail({ clubId });
-    return await this.clubMemberRepository.findWithPagination(paginationOptions, { club });
+    return await this.clubMemberRepository.findWithPagination(paginationOptions, { club }, { populate: ['user', 'club', 'club.members', 'club.members.user'] });
+  }
+
+  public async findClubMembership(
+    userId: string,
+    paginationOptions?: PaginationOptions,
+  ): Promise<PaginatedResult<ClubMember>> {
+    return await this.clubMemberRepository.findWithPagination(
+      paginationOptions,
+      { user: { userId } },
+      { populate: ['user', 'club', 'club.members'] },
+    );
   }
 
   public async addUserToClub(
+    requester: User,
     clubId: number,
     userId: string,
     createClubMemberDto: CreateClubMemberDto,
 ): Promise<ClubMember> {
-    const club = await this.clubRepository.findOneOrFail({ clubId });
+    const club = await this.clubRepository.findOneOrFail({ clubId }, ['members', 'members.user']);
+
+    // TODO: Move this to CASL
+    if (!requester.roles.includes(Role.Admin) && !club.isClubAdmin(requester))
+      throw new ForbiddenException('Not a club admin');
+
+    if (club.canActOnRole(requester, createClubMemberDto.role))
+      throw new BadRequestException('Role too high');
+
     const user = await this.userRepository.findOneOrFail({ userId });
     const existing = await this.clubMemberRepository.count({ club, user });
     if (existing)
@@ -79,25 +113,42 @@ export class ClubsService {
     return clubMember;
   }
 
-  public async findClubMembership(user: User): Promise<ClubMember[]> {
-    return await this.clubMemberRepository.findAll(user);
-  }
-
   public async updateUserRole(
+    requester: User,
     clubId: number,
     userId: string,
     updateClubMemberDto: UpdateClubMemberDto,
   ): Promise<ClubMember> {
-    const user = await this.userRepository.findOneOrFail({ userId });
-    const clubMember = await this.clubMemberRepository.findOneOrFail({ club: { clubId }, user });
+    const club = await this.clubRepository.findOneOrFail({ clubId }, ['members', 'members.user']);
+
+    // TODO: Move this to CASL
+    if (!requester.roles.includes(Role.Admin) && !club.isClubAdmin(requester))
+      throw new ForbiddenException('Not a club admin');
+
+    if (typeof updateClubMemberDto.role !== 'undefined' && club.canActOnRole(requester, updateClubMemberDto.role))
+      throw new BadRequestException('Role too high');
+
+    const clubMember = await this.clubMemberRepository.findOneOrFail({ club: { clubId }, user: { userId } }, ['user', 'club', 'club.members']);
 
     wrap(clubMember).assign(updateClubMemberDto);
     await this.clubMemberRepository.flush();
     return clubMember;
   }
 
-  public async removeUserFromClub(clubId: number, user: User): Promise<void> {
-    const clubMember = await this.clubMemberRepository.findOneOrFail({ club: { clubId }, user });
+  public async removeUserFromClub(requester: User, clubId: number, userId: string): Promise<void> {
+    const club = await this.clubRepository.findOneOrFail({ clubId }, ['members', 'members.user']);
+
+    const isSelf = requester.userId === userId;
+
+    // TODO: Move this to CASL
+    if (!isSelf && !requester.roles.includes(Role.Admin) && !club.isClubAdmin(requester))
+      throw new ForbiddenException('Not a club admin');
+
+    const clubMember = await this.clubMemberRepository.findOneOrFail({ club: { clubId }, user: { userId } });
+
+    if (!isSelf && clubMember.role === ClubRole.President)
+      throw new ForbiddenException('Cannot remove president');
+
     await this.clubMemberRepository.removeAndFlush(clubMember);
   }
 }
